Tune machines to obfuscate the circuit setup cell sequence.
I suggest you re-read the machines.c file instead of reading the diff here.

We abort the "look-like-nothing" strategy of the old machines, for these new
machines tha are carefully hand-tuned to obfuscate the circuit construction
sequence of intro/rend circuits.

We take this approach because it's easier to reason about the properties that
the machines offer, because they have minimal overhead, and also because they
are extensible by adding more states to them to do more advanced tricks