sendme: Off by one on the SENDME window
The proposal 289 specifies that the cell digest of the
(CIRCWINDOW_INCREMENT * i)-nth cell should be put in the SENDME and thus
validate on the other side.

We were off by one because originally I thought the SENDME would count in the
package window but it doesn't, only RELAY DATA cell do.

It doesn't break in functionnality, just that we were offsetted. Fortunately
for us, 0.4.1 is not released yet.

Part of #30428.

Signed-off-by: David Goulet <dgoulet@torproject.org>