Bug 30942: Make log a protocol warn.
I think it is a violation of the protocol to ignore invalid requests, and we
should find the most common of these in testing, to check them for validity.
Integration testing with the vanguards addon would help with this.
But that is a long term project...

If our testing infra would explode from frequent protocol warns, maybe omit
this commit. (But I don't think these will be that frequent).