Handle errors from execve() in the Unix process backend more gracefully.
This patch removes a call to tor_assert_unreached() after execve()
failed. This assertion leads to the child process emitting a stack trace
on its output, which makes the error harder for the user to demystify,
since they think it is an internal error in Tor instead of "just" being
a "no such file or directory" error.

See: https://bugs.torproject.org/31810