kist: Set scheduler interval to 2msec for clients
KIST cell scheduler is called if at least one cell is put on a circuit queue.
A fullscheduler run might not handle all cells because it depends on the
available space in the TCP buffer for the socket. What KIST does at the moment
is reschedule itself in 10ms (static value), the KISTSchedRunInterval
parameter. This is done so we can properly handle priority between circuits
with this grace period (mostly loud vs quiet circuits).

The problem is that if tor has very few sockets to handle, like most Tor
clients they have a single connection to their Guard, then it process that
what it can (usually in couple msec) and then sleep for the remaining time.

That remaining time would be between 8 and 9 msec doing nothing even though
bytes might be ready to be sent.

This caps a tor client to a maximum of ~3MB/sec (experimental results) instead
of being able to push much more if needed, let say a bulk download.

With this, we are bringing down the cell scheduler interval _only_ for clients
to 2msec. Relays stay on 10msec because they handle a large amount of sockets
and thus there is no starvation on that side.

This is really a bandaid to #29427 but there is a wider architectural problem
with that ticket so we are not closing it with this commit.

Part of #29427.

Signed-off-by: David Goulet <dgoulet@torproject.org>