test: Add setup/cleanup fn for test_scheduler.c
Because of the new EWMA touch function, the EWMA tick must be initialized. In
order to do that for each test we currently have and future ones we will
implement, this commit adds a setup and cleanup function for every tests in
test_scheduler.c in order to initialize the EWMA tick.

In order to do that, a bit of refactor is done to use a helper macro to setup
a test that always sets the testsetup_t object.

Finally, the loop_kist test is fixed to not reallocate the cmux because
new_fake_channel() does it in the first place and also sets the cmux policy
which we now need in ewma_touch().

Part of #29698

Signed-off-by: David Goulet <dgoulet@torproject.org>