Another attempt to fix Doxygen references in out-of-tree builds
This time, we're using cd to run doxygen from inside the source
directory, not the build directory at all.  This lets us call the
source directory ".", which (I hope) both Doxygen 1.8.13 and 1.8.15
will understand.

Naturally, this requires a corresponding change in the doxygen
configuration so that we are directing the output to the correct
place.

Fix for 32378.