Script to set us up for building all of our headers by themselves
This script makes a directory called "test_headers", then populates
it with one C file for each header.  The C file defines all of the
*_PRIVATE macros listed in the header, and then includes the header.

The script also makes a Makefile in this directory that tries to
build every C file it generates.

This script is not production-ready: it does not handle out-of-tree
operation, and doesn't look at the flags actually set by autoconf.