Always escape BridgeDistribution value from torrc in descriptors.
When the BridgeDistribution value was added in commit
ebab5215252f9fa810ae091c335c5ae6e619faaf the check for whether the value
contains '\n' or '\r' was added as a requirement for whether or not the
value should be escaped.

This patch removes this check and makes sure we call `escaped()` on the
value every time before we add it to a descriptor.

See: https://bugs.torproject.org/32753