dirlist: Add a function to test if an address is an authority
For now, and because of backport considerations, the patch is minimal as in it
iterates over the list of trusted directory authorities in order to match the
address. We do have very few entries in that list (currently 9 + 1) so the
cost in performance shouldn't be too bad.

If it shows up in the profile, we can move to an address set (_not_
address_set.c since that one is probabilistic) and thus perform lookup O(1).

Part of #33029

Signed-off-by: David Goulet <dgoulet@torproject.org>