dirauth: Resume sending 503 directory error code
Authorities were never sending back 503 error code since they are considered
to always be able to answer requests.

However, that recently backfired because of a large number of requests from
old clients using the DirPort that were _not_ getting their 503 code which
overloaded the DirPort leading to the authority unable to answer to its fellow
authorities.

This is not a complete solution to the problem but it will help ease off the
load on the authority side by sending back 503 codes *unless* the connection
is from a known relay or an authority.

To clarify further the situation: A new relay won't be known by an authority
so it might get a 503 code and thus fail the upload of its brand new
descriptor. However, it should be accepted by a majority of authorities and
thus its descriptor will end up in the consensus and fetched by the authority
that was unable to accept the upload request in the first place leading to the
relay to be recognized for future upload.

Fixes #33029

Signed-off-by: David Goulet <dgoulet@torproject.org>