Set subsystems_thread_cleanup as the thread cleanup callback.
Fixes bug 32103.

I'm calling this a bugfix on 0.4.1.1-alpha, since that is when we
introduced the first thread-cleanup code that actually does
anything.  (Previously, we had thread-cleanup code that called an
openssl per-thread cleanup function, but that function only existed
on older versions of openssl that we no longer support.)