hs-v3: Don't close intro circuit for an encrypted descriptor
To close intro circuits, we need a valid decrypted descriptor in order to find
the circuit by authentication key. However, we can have encrypted descriptors
in the client cache that are waiting for a client authorization token.

This commit doesn't attempt to close intro circuits for that situation because
we are unable to access the authentication key for each intro circuit.

Signed-off-by: David Goulet <dgoulet@torproject.org>